+++ test/cases/exception-pug-compile/src/index.js
@@ -1,0 +1,4 @@
+const tmpl = require('./index.pug');
+const html = tmpl();
+
+console.log(html);