+++ test/cases/javascript-option-method-rtrender/src/index.js
@@ -1,0 +1,4 @@
+// require with alias 'Template'
+const html = require('Template/widget.pug?{"a":10,"b":"abc"}');
+
+console.log(html);