+++ test/cases/javascript-option-method-rtrender/webpack.config.js
@@ -1,32 +1,32 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const basePath = path.resolve(__dirname);
-const webRootPath = path.join(__dirname, 'public/');
 
 module.exports = {
   mode: 'production',
-  entry: {},
+  entry: {
+    index: './src/index.js',
+  },
   resolve: {
     alias: {
       Includes: path.join(basePath, 'src/includes/'),
+      Images: path.join(basePath, 'src/images/'),
       Template: path.join(basePath, 'src/template/'),
     },
   },
-  plugins: [
-    new HtmlWebpackPlugin({
-      filename: path.join(webRootPath, '/index.html'),
-      template: './src/template/index.pug',
-      inject: false,
-    }),
-  ],
+  plugins: [],
   module: {
     rules: [
       {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          method: 'compile',
+          method: 'rtRender',
         },
+      },
+      // process image resources in pug templates
+      {
+        test: /\.(png|jpg|jpeg)/,
+        type: 'asset/resource',
       },
     ],
   },